+++ src/_assets/js/app.js
@@ -1,5 +1,7 @@
 // JS
 const browser = require( './browser' );
+const a11y = require( './a11y' );
+const navigation = require( './navigation' );
 
 // SCSS
 import './../css/app.css';