+++ src/_assets/js/development.js
@@ -1,0 +1,9 @@
+// Load development CSS
+import './../css/development.css';
+
+if ( process.env.NODE_ENV == 'development' ) {
+
+	// Log environment to the console
+	console.log( process.env.NODE_ENV + ' environment');
+
+}