+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Install and configure Eleventy
+- Add `.eleventy.js`
 
 ### Changed
 - Update npm scripts