+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Install and configure Eleventy
 - Add `.eleventy.js`
+- Add documentation in `_docs` directory
 
 ### Changed
 - Update npm scripts